Moved some demo code to App.jsx from Canvas.jsx
